+++ docs/changelog.rst
@@ -22,6 +22,9 @@ Feature release.
 - Core: Fix crash on `library.lookup(uris=[])`. (Fixes: :issue:`1619`, PR:
   :issue:`1620`)
 
+- Core: Define return value of `playlists.delete()` to be a bool, :class:`True`
+  on success, :class:`False` otherwise. (PR: :issue:`1702`)
+
 - File: Change default ordering to show directories first, then files. (PR:
   :issue:`1595`)